# specs/001-stock-ai-decision-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 股票 AI 策略回测决策系统 API
  description: 基于多个 AI 策略回测模型的股票交易决策系统
  version: 1.0.0
  contact:
    name: API Support
    email: support@stock-ai.com
servers:
  - url: http://localhost:8099/api/v1
    description: 开发环境
  - url: https://api.stock-ai.com/api/v1
    description: 生产环境
paths:
...
  /decisions/recommendations:
    get:
      summary: 获取股票推荐
      description: 获取基于多模型投票的股票推荐列表
      tags:
        - Decisions
      parameters:
        - name: industry
          in: query
          description: 行业筛选
          schema:
            type: string
        - name: limit
          in: query
          description: 推荐数量
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: 推荐股票列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
  /decisions/{decision_id}:
    get:
      summary: 获取决策详情
      description: 获取指定决策的详细信息，包括投票详情
      tags:
        - Decisions
      parameters:
        - name: decision_id
          in: path
          required: true
          description: 决策ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 决策详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionDetailResponse"
        "404":
          description: 决策不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # 模型相关 API
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        data:
          type: object
          description: 响应数据
        message:
          type: string
          description: 响应消息
        status:
          type: string
          description: 状态码
          enum: [success, error]
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            error_code:
              type: string
              description: 错误代码
            details:
              type: object
              description: 错误详情
...
    Stock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        symbol:
          type: string
        name:
          type: string
        industry:
          type: string
        current_price:
          type: number
          format: float
        price_change:
          type: number
          format: float
        price_change_percent:
          type: number
          format: float
        volume:
          type: integer
        market_cap:
          type: number
          format: float
        pe_ratio:
          type: number
          format: float
        pb_ratio:
          type: number
          format: float
...
    Decision:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stock:
          $ref: "#/components/schemas/Stock"
        decision_type:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        target_price:
          type: number
          format: float
        stop_loss_price:
          type: number
          format: float
        time_horizon:
          type: integer
        reasoning:
          type: string
        generated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...
    VoteResult:
      type: object
      properties:
        model_id:
          type: string
          format: uuid
        model_name:
          type: string
        vote_type:
          type: string
          enum: [buy, sell, hold]
        confidence:
          type: number
          format: float
        signal_strength:
          type: number
          format: float
        reasoning:
          type: string
    DecisionDetailResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                decision:
                  $ref: "#/components/schemas/Decision"
                vote_results:
                  type: array
                  items:
                    $ref: "#/components/schemas/VoteResult"
    RecommendationResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              type: object
              properties:
                recommendations:
                  type: array
                  items:
                    $ref: "#/components/schemas/Decision"

    # 模型相关模型
...